server-main/pyproject.toml
[tool.poetry]
name = "server_main"
version = "0.1.0"
description = ""
authors = ["BoolmanO <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
alembic = "^1.13.2"
pydantic-settings = "^2.3.4"
yarl = "^1.9.4"
pyjwt = "^2.8.0"
cryptography = "^43.0.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
loguru = "^0.7.2"
redis = {extras = ["hiredis"], version = "^5.0.7"}
aio-pika = "^9.4.2"
result = "^0.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
